creating the view by profiles tables in the view.
The view should be working but is not. Details are as follows:

Results from server are correct and can be verified by a print statement found in dispatcher.
The html with the results injected into it which is appended to the "view by profiles" Div is correct and can be found in the console.log in the web browser (ctrl + shift + J	in Chrome). When this html is copied into main_1.html for testing it works. However when it is generated in the main.jsp it does not render anything more than the inner "profile x" tabs. This is not because it is .jsp rather than .html as I've already tried converting the webpage to .html with no success.

I am uncertain as to the cause of this problem, however a hint toward its natures is that the url appends the name of the inner "profile x" tab when that tab is clicked. This does NOT happen in the main_1.html test version of the webpage.